requirements.txt
aergo-herapy==2.0.1
PyInquirer
pyfiglet